+++ tests/ui/builder-method-with-skip-setter.rs
@@ -1,0 +1,9 @@
+use type_state_builder::TypeStateBuilder;
+
+#[derive(TypeStateBuilder)]
+struct Invalid {
+    #[builder(required, builder_method, skip_setter)]
+    id: u64,
+}
+
+fn main() {}